% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_function.R
\name{check_inputs_compute_gini}
\alias{check_inputs_compute_gini}
\title{check_inputs_compute_gini}
\usage{
check_inputs_compute_gini(df, welfare_col, weight_col)
}
\arguments{
\item{df}{data.frame: data.frame input to compute_gini()}

\item{welfare_col}{character: Name of welfare column
For instance, income.}

\item{weight_col}{character: Name of weight column. I}
}
\value{
logical
}
\description{
check_inputs_compute_gini
}
