% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_function.R
\name{compute_gini}
\alias{compute_gini}
\title{compute_gini
Computes gini coefficient}
\usage{
compute_gini(df, welfare_col = "welfare", weight_col = "weight")
}
\arguments{
\item{df}{data.frame: Data frame containing variables for the gini computation}

\item{welfare_col}{character: Name of welfare column
For instance, income.}

\item{weight_col}{character: Name of weight column. In case each observation
represents a different proportion of the overall population}
}
\value{
numeric
}
\description{
compute_gini
Computes gini coefficient
}
\examples{
welfare <- c(1, 5, 100)
weight <- c(.5, .3, .2)
df <- data.frame(welfare = welfare, weight = weight)
compute_gini(df)
}
